docs: move detailed Git/PR workflow to dedicated contribution guide (#2294)
